public

+++ routes/logout.go
@@ -12,7 +12,7 @@ import (
 func LogoutRouter(c echo.Context) error {
 	session, _ := session.Get("oursession", c)
 	session.Options.MaxAge = -1 // removes cookie
-	session.Values["userid"] = uint(0)
+	session.Values["userid"] = 0
 	_ = session.Save(c.Request(), c.Response())
 	c.Set("UserID", 0)
 	return c.Redirect(http.StatusFound, "/login")
